cpp11 build and test

See merge request ifs/coast/coast!1

+++ coast/modules/StdDataAccess/Test/mkca-new.sh
@@ -7,26 +7,25 @@
 # the license that is included with this library/application in the file license.txt.
 #--------------------------------------------------------------------
 
-MYNAME=`basename $0`
+MYNAME=$(basename "$0")
 
 country="CH"
 state="ZH"
 locality="ZH"
-org="Test"
 orgunit="CertIssueing"
 emailsuffix="wildplanet.com"
 commonnameclient=""
 commonnameserver=""
 pathtoexistingcacert=""
 digest="sha1"
-openssl_bin=`which openssl`
+openssl_bin=$(command -v openssl)
 openssl_binOpts=""
 conf=/etc/ssl/openssl.cnf
 target_dir=""
-serialnum=`date +'%s'`
+serialnum=$(date +"%s")
 if [ "$serialnum" = "%s" ]; then
 	# date seems to be non-gnu style, workaround approximation!
-	serialnum=`expr \( \( \`date +%Y\` - 1970 \) \* 365 + \( \`date +%Y\` - 1970 \) / 4 - 1 + \`date +%j\` \) \* 86400 + \`date +%H\` \* 3600 + \`date +%M\` \* 60 + \`date +%S\` `
+	serialnum=$(( ( ( $(date +%Y) - 1970 ) * 365 + ( $(date +%Y) - 1970 ) / 4 - 1 + $(date +%j) ) * 86400 + $(date +%H) * 3600 + $(date +%M) * 60 + $(date +%S) ))
 fi
 
 errorexit()
@@ -37,20 +36,22 @@ errorexit()
 
 showhelp()
 {
-	echo ''
-	echo 'usage: '$MYNAME' [options]'
-	echo 'where options are:'
-	echo ' -b <path2sslbin> : full path and name to openssl binary, default '${openssl_bin}
-	echo ' -B <options>     : options to openssl binary given in -b'
-	echo ' -o <path2sslcnf> : full path and name to ssl configuration, default '${conf}
-    echo ' -d <path>        : directory to place CA-Structure into. creates ca-hierarchy within this directory'
-    echo ' -D <days>        : validity duration of generated certificates, default is 3650'
-    echo ' -c <path>        : copy the created certs to given location'
-    echo ' -E <path>        : location where previously built rootCA1 and subCA1 certs are kept'
-    echo '                    This allows you to add additional client/server certs without re-generating previously generated certificates.'
-    echo ' -C <common name> : subject of client, entity which is checked in cert verify, may be given repeatedly'
-    echo ' -S <common name> : subject of server, entity which is checked in cert verify, may be given repeatedly'
-    echo ' -s <digest>      : message digest used to sign server/client certificate. choose one of: md2/md5/sha1/mdc/mdc2, default is md5'
+	[ -n "${1}" ] && echo "${1}";
+	echo ""
+	echo "usage: $MYNAME [options]"
+	echo "where options are:"
+	echo " -b <path2sslbin> : full path and name to openssl binary, default ${openssl_bin}"
+	echo " -B <options>     : options to openssl binary given in -b"
+	echo " -o <path2sslcnf> : full path and name to ssl configuration, default ${conf}"
+    echo " -d <path>        : directory to place CA-Structure into. creates ca-hierarchy within this directory"
+    echo " -D <days>        : validity duration of generated certificates, default is 3650"
+    echo " -c <path>        : copy the created certs to given location"
+    echo " -E <path>        : location where previously built rootCA1 and subCA1 certs are kept"
+    echo "                    This allows you to add additional client/server certs without re-generating previously generated certificates."
+    echo " -C <common name> : subject of client, entity which is checked in cert verify, may be given repeatedly"
+    echo " -S <common name> : subject of server, entity which is checked in cert verify, may be given repeatedly"
+    echo " -s <digest>      : message digest used to sign server/client certificate. choose one of: md2/md5/sha1/mdc/mdc2, default is $digest"
+    echo " -e <emailsuffix> : email suffix to use, default is $emailsuffix"
     echo ''
     exit 3
 }
@@ -61,11 +62,10 @@ while getopts ":b:B:o:d:D:c:C:e:E:s:S:-" opt
 do
     case $opt in
 		:)
-			echo "ERROR: -$OPTARG parameter missing, exiting!";
-			showhelp;
+			showhelp "ERROR: -$OPTARG parameter missing, exiting!";
 		;;
         b)
-        	if [ -r ${OPTARG} -a -x ${OPTARG} ]; then
+        	if [ -r "${OPTARG}" ] && [ -x "${OPTARG}" ]; then
         		openssl_bin="${OPTARG}";
 			else
 				echo "openssl binary [${OPTARG}] not found or not executable, using default: ${openssl_bin}";
@@ -78,10 +78,10 @@ do
         ;;
         o)
     		# openssl needs absolute path to config file
-    		_conf="`dirname ${OPTARG}`";
-    		_conf="`cd ${_conf} 2>/dev/null && pwd -P`";
-    		_conf="${_conf}/`basename ${OPTARG}`";
-        	if [ -n "${_conf}" -a -r ${_conf} ]; then
+    		_conf="$(dirname "${OPTARG}")";
+    		_conf="$(cd "${_conf}" 2>/dev/null && pwd -P)";
+    		_conf="${_conf}"/$(basename "${OPTARG}");
+        	if [ -n "${_conf}" ] && [ -r "${_conf}" ]; then
 				conf="${_conf}";
     		fi;
         ;;
@@ -97,7 +97,7 @@ do
         ;;
         S)      commonnameserver="${commonnameserver} ${OPTARG}";
         ;;
-        e)      emailclient="${OPTARG}";
+        e)      emailsuffix="${OPTARG}";
         ;;
         E)      pathtoexistingcacert="${OPTARG}";
         ;;
@@ -110,20 +110,18 @@ do
     esac
 done
 
-shift `expr $OPTIND - 1`
+shift $((OPTIND - 1))
 
-if [ ! -x ${openssl_bin} ]
+if [ ! -x "${openssl_bin}" ]
 then
-	echo "Could not find openssl binary."
-	showhelp;
+	showhelp "Could not find openssl binary."
 fi
 echo "Using openssl binary ${openssl_bin}"
 openssl_bin="${openssl_bin} ${openssl_binOpts}"
 
-if [ ! -f $conf ]
+if [ ! -f "$conf" ]
 then
-    echo "Could not find openssl configuration file ${conf}"
-    showhelp;
+    showhelp "Could not find openssl configuration file ${conf}"
 fi
 echo "Using openssl config file ${conf}"
 
@@ -138,26 +136,23 @@ fi
 
 if [ -z "${target_dir}" ]
 then
-	echo "target directory not set, exiting!"
-    showhelp;
+	showhelp "target directory not set, exiting!"
 fi
 
-target_dir=${target_dir}/ca-hierarchy
+target_dir="${target_dir:?}"/ca-hierarchy
 
-if [ ! -d ${target_dir} ]
+if [ ! -d "${target_dir}" ]
 then
-    mkdir -p ${target_dir}
-    if [ $? -ne 0 ]
-    then
+    if ! mkdir -p "${target_dir:?}"; then
         echo "Can't create ${target_dir}"
         exit 1
     fi
 fi
-cd ${target_dir}
+cd "${target_dir:?}" || { "Cannot change to directory ${target_dir}, exiting!"; exit 4; }
 
 echo "Creating RootCA1"
-rm -Rf ${target_dir}/*
-mkdir ${target_dir}/ucerts
+rm -Rf "${target_dir:?}"/*
+mkdir "${target_dir:?}"/ucerts
 
 CATOP=demoCA
 CAKEY=cakey.pem
@@ -174,13 +169,13 @@ touch ${CATOP}/index.txt
 if [ -z "${pathtoexistingcacert}" ]
 then
 	${openssl_bin} req \
-		-config $conf \
+		-config "$conf" \
 		-new \
 		-x509 \
 		-passout pass:hank \
 		-keyout ${CATOP}/private/$CAKEY \
 		-out ${CATOP}/$CACERT \
-		-days $days <<-EOF
+		-days "$days" <<-EOF
 			${country}
 			${state}
 			${locality}
@@ -204,15 +199,15 @@ echo "Stripping passphrase from rootCA1key.pem"
 ${openssl_bin} rsa \
 	-passin pass:hank \
 	-in ${CATOP}/private/$CAKEY \
-	-out ${target_dir}/ucerts/rootCA1key.pem
+	-out "${target_dir:?}"/ucerts/rootCA1key.pem
 
 echo "Certificate Request for SubCA1"
 ${openssl_bin} req \
-	-config $conf \
+	-config "$conf" \
 	-new \
 	-passout pass:gugus \
-	-keyout ${target_dir}/newreq.pem \
-	-out ${target_dir}/newreq.pem <<-EOF1
+	-keyout "${target_dir:?}"/newreq.pem \
+	-out "${target_dir:?}"/newreq.pem <<-EOF1
 		${country}
 		${state}
 		${locality}
@@ -226,19 +221,19 @@ EOF1
 
 echo "Signing CA request for SubCA1"
 ${openssl_bin} ca \
-	-config $conf \
+	-config "$conf" \
 	-policy policy_anything \
 	-passin pass:hank \
-	-out ${target_dir}/newcert.pem \
+	-out "${target_dir:?}"/newcert.pem \
 	-extensions v3_ca \
-	-days $days \
-	-infiles ${target_dir}/newreq.pem <<-EOF2
+	-days "$days" \
+	-infiles "${target_dir:?}"/newreq.pem <<-EOF2
 		y
 		y
 EOF2
 else
-	cp ${pathtoexistingcacert}/rootCA1key.pem ${CATOP}/private/$CAKEY
-	cp ${pathtoexistingcacert}/rootCA1crt.pem ${CATOP}/$CACERT
+	cp "${pathtoexistingcacert:?}"/rootCA1key.pem ${CATOP}/private/$CAKEY
+	cp "${pathtoexistingcacert:?}"/rootCA1crt.pem ${CATOP}/$CACERT
 	if [ ! -f ${CATOP}/private/$CAKEY ]
 	then
 		echo "Unable to retrieve rootCA1key.pem"
@@ -256,7 +251,7 @@ fi
 echo "Creating SubCA1"
 
 mkdir subCA1
-cd subCA1
+cd subCA1 || { echo "Failed to create subCA1, exiting!"; exit 5; }
 
 mkdir ${CATOP}
 mkdir ${CATOP}/certs
@@ -269,22 +264,22 @@ touch ${CATOP}/index.txt
 if [ -z "${pathtoexistingcacert}" ]
 then
 	echo "Stripping Certificate and private Key from Certificate for SubCA1"
-	csplit -f work -s -n 2  ${target_dir}/newreq.pem '/-----BEGIN CERTIFICATE REQUEST-----/'
+	csplit -f work -s -n 2  "${target_dir:?}"/newreq.pem '/-----BEGIN CERTIFICATE REQUEST-----/'
 	mv work00 ${CATOP}/private/cakey.pem
 	rm work01
 	echo "Stripping passphrase from subCA1key.pem"
 	${openssl_bin} rsa \
 		-passin pass:gugus \
 		-in ${CATOP}/private/$CAKEY \
-		-out ${target_dir}/ucerts/subCA1key.pem
-	cp ${CATOP}/private/cakey.pem ${target_dir}/ucerts/subCA1key.pem
+		-out "${target_dir:?}"/ucerts/subCA1key.pem
+	cp ${CATOP}/private/cakey.pem "${target_dir:?}"/ucerts/subCA1key.pem
 	echo "Stripping plaintext from SubCA1 Certificate"
 	${openssl_bin} x509 \
-		-in ${target_dir}/newcert.pem \
+		-in "${target_dir:?}"/newcert.pem \
 		-out ${CATOP}/cacert.pem
 else
-	cp ${pathtoexistingcacert}/subCA1key.pem ${CATOP}/private/$CAKEY
-	cp ${pathtoexistingcacert}/subCA1crt.pem ${CATOP}/cacert.pem
+	cp "${pathtoexistingcacert:?}"/subCA1key.pem ${CATOP}/private/$CAKEY
+	cp "${pathtoexistingcacert:?}"/subCA1crt.pem ${CATOP}/cacert.pem
 	if [ ! -f ${CATOP}/private/$CAKEY ]
 	then
 		echo "Unable to retrieve subCA1key.pem"
@@ -303,7 +298,7 @@ for commonnameserver_var in ${commonnameserver}
 do
 	echo "Creating certificate request for Server ${commonnameserver_var}"
 	${openssl_bin} req \
-		-config $conf \
+		-config "$conf" \
 		-new \
 		-passout pass:gaga \
 		-keyout newreq_${commonnameserver_var}.pem \
@@ -321,11 +316,11 @@ EOF1
 
 	echo "Signing CA request for Server ${commonnameserver_var}"
 	${openssl_bin} ca \
-		-md $digest \
-		-config $conf \
+		-md "$digest" \
+		-config "$conf" \
 		-policy policy_anything \
 		-passin pass:gugus \
-		-days $days \
+		-days "$days" \
 		-out newcert_${commonnameserver_var}.pem \
 		-infiles newreq_${commonnameserver_var}.pem <<-EOF2
 			y
@@ -343,7 +338,7 @@ EOF2
 		-out newreq_${commonnameserver_var}.plain.pem
 done
 
-mkdir ${target_dir}/client && cd ${target_dir}/client && (
+mkdir "${target_dir:?}"/client && cd "${target_dir:?}"/client && (
 
 	for commonnameclient_var in ${commonnameclient}
 	do
@@ -353,7 +348,7 @@ mkdir ${target_dir}/client && cd ${target_dir}/client && (
 
 		echo "Creating certificate request for Client ${commonnameclient_var}"
 		${openssl_bin} req \
-			-config $conf \
+			-config "$conf" \
 			-new \
 			-passout pass:gaga \
 			-key client_${commonnameclient_var}.key \
@@ -371,12 +366,12 @@ mkdir ${target_dir}/client && cd ${target_dir}/client && (
 
 		echo "Signing Client certificte for ${commonnameclient_var} by subCA1"
 		${openssl_bin} x509 \
-			-$digest \
+			-"$digest" \
 			-req \
-			-days $days \
+			-days "$days" \
 			-passin pass:gugus \
-			-CA ${target_dir}/subCA1/${CATOP}/cacert.pem \
-			-CAkey ${target_dir}/subCA1/${CATOP}/private/cakey.pem \
+			-CA "${target_dir:?}"/subCA1/${CATOP}/cacert.pem \
+			-CAkey "${target_dir:?}"/subCA1/${CATOP}/private/cakey.pem \
 			-CAcreateserial \
 			-in client_${commonnameclient_var}.csr \
 			-out clientcrt_${commonnameclient_var}.pem
@@ -393,34 +388,34 @@ mkdir ${target_dir}/client && cd ${target_dir}/client && (
 	done
 )
 
-cd ${target_dir}/subCA1 && (
-	echo "Building result directory  ${target_dir}/ucerts"
-	cp  ${CATOP}/cacert.pem ${target_dir}/ucerts/subCA1crt.pem
+cd "${target_dir:?}"/subCA1 && (
+	echo "Building result directory ${target_dir:?}/ucerts"
+	cp  ${CATOP}/cacert.pem "${target_dir:?}"/ucerts/subCA1crt.pem
 	for commonnameserver_var in ${commonnameserver}
 	do
-		cp  newcert_${commonnameserver_var}.pem       ${target_dir}/ucerts/servercrt_${commonnameserver_var}.pem
-		cp  newreq_${commonnameserver_var}.plain.pem  ${target_dir}/ucerts/serverkey_${commonnameserver_var}.pem
+		cp  newcert_${commonnameserver_var}.pem       "${target_dir:?}"/ucerts/servercrt_${commonnameserver_var}.pem
+		cp  newreq_${commonnameserver_var}.plain.pem  "${target_dir:?}"/ucerts/serverkey_${commonnameserver_var}.pem
 	done
 )
 
-cd ${target_dir} && (
-	cp ${CATOP}/cacert.pem ${target_dir}/ucerts/rootCA1crt.pem
+cd "${target_dir:?}" && (
+	cp ${CATOP}/cacert.pem "${target_dir:?}"/ucerts/rootCA1crt.pem
 	if [ -z "${pathtoexistingcacert}" ]
 	then
-		cp ${CATOP}/cacert.p12 ${target_dir}/ucerts/rootCA1.p12
+		cp ${CATOP}/cacert.p12 "${target_dir:?}"/ucerts/rootCA1.p12
 	fi
 )
 
-cd ${target_dir}/client && (
+cd "${target_dir:?}"/client && (
 	for commonnameclient_var in ${commonnameclient}
 	do
-		cp  clientcrt_${commonnameclient_var}.p12      ${target_dir}/ucerts/clientcrt_${commonnameclient_var}.p12
-		cp  clientcrt_${commonnameclient_var}.pem	   ${target_dir}/ucerts/clientcrt_${commonnameclient_var}.pem
-		cp  client_${commonnameclient_var}.key		   ${target_dir}/ucerts/clientkey_${commonnameclient_var}.pem
+		cp  clientcrt_${commonnameclient_var}.p12      "${target_dir:?}"/ucerts/clientcrt_${commonnameclient_var}.p12
+		cp  clientcrt_${commonnameclient_var}.pem	   "${target_dir:?}"/ucerts/clientcrt_${commonnameclient_var}.pem
+		cp  client_${commonnameclient_var}.key		   "${target_dir:?}"/ucerts/clientkey_${commonnameclient_var}.pem
 	done
 )
 
-cd ${target_dir}/ucerts && (
+cd "${target_dir:?}"/ucerts && (
 	echo "Creating Chain without Server Cert)"
 	cp  subCA1crt.pem     chainwoservercrt.pem
 	cat rootCA1crt.pem >> chainwoservercrt.pem
@@ -441,39 +436,39 @@ cd ${target_dir}/ucerts && (
 	done
 )
 
-echo "Your certificates are in ${target_dir}/ucerts"
-if [ ! -z "${copyto}" ]
+echo "Your certificates are in ${target_dir:?}/ucerts"
+if [ -n "${copyto}" ]
 then
-	if [ ! -d ${copyto} ]
+	if [ ! -d "${copyto}" ]
 	then
-		mkdir ${copyto}
+		mkdir "${copyto:?}"
 	fi
 	echo "Copying SERVER related files: serverkey servercrt ...  to ${copyto}:"
-	cd ${target_dir}/ucerts && (
+	cd "${target_dir:?}"/ucerts && (
 		for commonnameserver_var in ${commonnameserver}
 		do
-			cp servercrt_${commonnameserver_var}.pem          ${copyto}
-			cp serverkey_${commonnameserver_var}.pem          ${copyto}
-			cp chainwoservercrt.pem   ${copyto}
-			cp fullchain_${commonnameserver_var}.pem          ${copyto}
-			cp serverchain_${commonnameserver_var}.pem    	  ${copyto}
+			cp servercrt_${commonnameserver_var}.pem          "${copyto}"
+			cp serverkey_${commonnameserver_var}.pem          "${copyto}"
+			cp chainwoservercrt.pem   "${copyto}"
+			cp fullchain_${commonnameserver_var}.pem          "${copyto}"
+			cp serverchain_${commonnameserver_var}.pem    	  "${copyto}"
 		done
 		echo "Copying CLIENT related files: clientcrt clientkey clientcrt.p12 to ${copyto}:"
 		for commonnameclient_var in ${commonnameclient}
 		do
-			cp clientcrt_${commonnameclient_var}.pem	 ${copyto}
-			cp clientkey_${commonnameclient_var}.pem	 ${copyto}
-			cp clientcrt_${commonnameclient_var}.p12	 ${copyto}
+			cp clientcrt_${commonnameclient_var}.pem	 "${copyto}"
+			cp clientkey_${commonnameclient_var}.pem	 "${copyto}"
+			cp clientcrt_${commonnameclient_var}.p12	 "${copyto}"
 		done
 
 		echo "Copying CA related files: subCA1 rootCA1 to ${copyto}:"
-		cp subCA1*.pem	 ${copyto}
-		cp rootCA1*.pem	 ${copyto}
+		cp subCA1*.pem	 "${copyto}"
+		cp rootCA1*.pem	 "${copyto}"
 	)
 fi
 
-echo "Verifying certificates in ${target_dir}/ucerts"
-cd ${target_dir}/ucerts
+echo "Verifying certificates in ${target_dir:?}/ucerts"
+cd "${target_dir:?}"/ucerts || { echo "Failed to change directory to ${target_dir:?}/ucerts, exiting!"; exit 6; }
 for commonnameserver_var in ${commonnameserver}
 do
 	echo "Verifying cert chain with -untrusted option for ${commonnameserver_var}"
